requirements.txt
requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1